% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interval.R
\name{check_interval}
\alias{check_interval}
\title{Check chr, start_bp, end_bp for validity}
\usage{
check_interval(chr, start_bp, end_bp)
}
\arguments{
\item{chr}{chromosome number}

\item{start_bp}{starting position in bp}

\item{end_bp}{ending position in bp}
}
\value{
none
}
\description{
Simple checks.
}

