% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interval.R
\name{convert_bp}
\alias{convert_bp}
\title{Convert to bp if in Mb}
\usage{
convert_bp(value, to_bp = TRUE)
}
\arguments{
\item{value}{value in bp or Mbp}

\item{to_bp}{convert to bp if TRUE, Mbp if FALSE (default TRUE)}
}
\value{
value in bp
}
\description{
Assumes positive value.
}
