% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_snp.R
\name{plot.feature_snp}
\alias{plot.feature_snp}
\alias{summary.feature_snp}
\title{Plot of features with SNPs}
\usage{
\method{plot}{feature_snp}(x, ...)

\method{summary}{feature_snp}(object, ...)
}
\arguments{
\item{x}{tbl of feature information from \code{\link{get_mgi_features}}}

\item{...}{arguments passed along to \code{\link{gene_plot}}}

\item{object}{tbl of feature information from \code{\link{get_feature_snp}}}

\item{...}{additional parameters ignored}
}
\value{
ggplot (see \code{\link{gene_plot}})

tbl of feature summaries by type
}
\description{
Uses \code{\link{gene_plot}} to plot genes, exons, mRNA with SNPs.

Summary of features with SNP information
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
\keyword{utilities}
