% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_tbl.R
\name{summary.feature_tbl}
\alias{summary.feature_tbl}
\alias{subset.feature_tbl}
\title{Summary of features}
\usage{
\method{summary}{feature_tbl}(object, major = TRUE)

\method{subset}{feature_tbl}(x, start_val = 0, stop_val = max(x$stop), ...)
}
\arguments{
\item{object}{tbl of feature information from \code{\link{get_mgi_features}}}

\item{major}{if \code{TRUE} (default), only summarize genes and exons}

\item{x}{tbl of feature information from \code{\link{get_feature_tbl}}}

\item{start_val, stop_val}{start and stop positions for subset}

\item{...}{additional parameters ignored}
}
\value{
tbl of feature summaries by type

tbl of feature summaries by type
}
\description{
Show count min and max of features by type

Subset of features
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
