% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exon.R
\name{get_gene_exon_snp}
\alias{get_gene_exon_snp}
\alias{get_gene_exon}
\alias{summary.gene_exon}
\alias{subset.gene_exon}
\title{Get exons for set of genes}
\usage{
get_gene_exon_snp(top_snps_tbl)

get_gene_exon(feature_tbl, gene_snp)

\method{summary}{gene_exon}(gene_exon, gene_name = NULL,
  top_snps_tbl = NULL, extra_bp = 5000)

\method{subset}{gene_exon}(x, gene_val, ...)
}
\arguments{
\item{top_snps_tbl}{table from \code{\link[qtl2scan]{top_snps}}}

\item{feature_tbl}{tbl of features from \code{query_variants}; see package \code{qtl2db}}

\item{gene_snp}{tbl of genes with SNPs IDs from \code{\link{match_feature_snp}}}

\item{gene_exon}{tbl of feature information from \code{\link{get_gene_exon}}}

\item{gene_name}{name of gene as character string}

\item{extra_bp}{extra region beyond gene for SNPs (in bp)}

\item{top_snps_tbl}{table of top SNPs in region from \code{\link[qtl2scan]{top_snps}}}
}
\value{
tbl of exon and gene features

tbl of exon and gene features

tbl of summary
}
\description{
Match up exon start,stop,strand with genes. Use \code{query_genes} to find features; see \code{qtl2db}.

Match up exon start,stop,strand with genes.

Returns table of gene and its exons.
}
\examples{
\dontrun{get_gene_exon_snp(top_snps_tbl)}

\dontrun{get_gene_exon(feature_snp)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
