% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_snp.R
\name{summary.gene_snp}
\alias{summary.gene_snp}
\title{Summary of genes overlapping SNPs}
\usage{
\method{summary}{gene_snp}(object, ...)
}
\arguments{
\item{object}{tbl of feature information from \code{\link{get_feature_snp}}}

\item{...}{additional parameters ignored}
}
\value{
tbl of feature summaries by type
}
\description{
Summary of genes overlapping SNPs
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
