% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_snp.R
\name{get_feature_snp}
\alias{get_feature_snp}
\title{Match features with SNPs}
\usage{
get_feature_snp(snp_tbl, feature_tbl, extend = 5000)
}
\arguments{
\item{snp_tbl}{tbl of SNPs from \code{assoc.map}}

\item{feature_tbl}{tbl of feature information from \code{\link{get_mgi_features}}}

\item{extend}{extend region for SNPs in bp (default 5000)}
}
\value{
tbl of features covering SNPs
}
\description{
Find features that overlap with SNPs
}
\examples{
\dontrun{get_feature_snp(snp_tbl, feature_tbl)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
