% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene.R
\name{get_gene}
\alias{get_gene}
\title{Get gene in region}
\usage{
get_gene(chr_id, start, stop)
}
\arguments{
\item{chr_id}{chromosome identifier}

\item{start}{start position in Mbp}

\item{stop}{stop position in Mbp}
}
\description{
Get gene using \code{query_genes}; see \code{qtl2db}.
}
