% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_snp.R
\name{get_gene_snp}
\alias{get_gene_snp}
\title{Match genes with SNPs}
\usage{
get_gene_snp(snp_tbl, feature_tbl, feature_snp = get_feature_snp(snp_tbl,
  feature_tbl, 0))
}
\arguments{
\item{snp_tbl}{tbl of SNPs from \code{assoc.map}}

\item{feature_tbl}{tbl of feature information from \code{query_variants}; see package \code{qtl2db}}

\item{feature_snp}{tbl of feature information from \code{\link{get_feature_snp}}}
}
\value{
tbl of genes covering SNPs
}
\description{
Find features that overlap with SNPs
}
\examples{
\dontrun{get_gene_snp(snp_tbl, feature_tbl)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
