% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mgi_features.R
\name{get_mgi_features}
\alias{get_mgi_features}
\title{Pull MGI gene tbl from SQLite database}
\usage{
get_mgi_features(chr_id, start_bp, end_bp, with_name = TRUE,
  gene_tbl = "mgi_gene", sql_file = "mgi_db.sqlite")
}
\arguments{
\item{chr_id}{chromosome number}

\item{start_bp}{starting position in bp}

\item{end_bp}{ending position in bp}

\item{with_name}{only pull named features if TRUE (default)}

\item{gene_tbl}{character string of tbl to be used in database}

\item{sql_file}{character string name of SQLite file (default "mgi_db.sqlite")}
}
\value{
tbl from SQLite database
}
\description{
Get subset of MGI gene table from SQLite using chromosome interval.
This is being replaced by \code{query_variants}. See package qtl2db.
}
\examples{
dontrun(get_mgi_features(9, 104*1e6, 109*1e6))

}
