% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpinfo.R
\name{get_snpinfo}
\alias{get_snpinfo}
\title{Get SNP and InDel information in window around peak}
\usage{
get_snpinfo(chr_id, peak_Mbp, window_Mbp, datapath, info_type = c("snps",
  "indels"))
}
\arguments{
\item{chr_id}{chromosome identifier}

\item{peak_Mbp}{position in Mbp of peak}

\item{window_Mbp}{half-width of \code{window} around \code{peak_Mbp}}

\item{datapath}{path to Derived Data}

\item{info_type}{type of info: one of c("snps","indels")}
}
\value{
table of class \code{snpinfo} of top_snps near maximum lod
}
\description{
Get SNP and InDel information from SQLite databases within \code{window_Mbp} of \code{peak_Mbp} on \code{chri_id}
}
\examples{
\dontrun{get_snpinfo(chr_id, peak_Mbp, window_Mbp)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
