% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpinfo.R
\name{get_snpprobs}
\alias{get_snpprobs}
\title{Get SNP genotype probabilities in window around peak}
\usage{
get_snpprobs(chr_id = NULL, peak_Mbp = NULL, window_Mbp = NULL, phename,
  probs_obj, datapath)
}
\arguments{
\item{chr_id}{chromosome identifier}

\item{peak_Mbp}{position in Mbp of peak}

\item{window_Mbp}{half-width of \code{window} around \code{peak_Mbp}}

\item{phename}{names of phenotypes}

\item{probs_obj}{object of class \code{\link[qtl2geno]{calc_genoprob}} for \code{chr_id}}

\item{datapath}{path to Derived Data}
}
\value{
table of class \code{near_snps} of top_snps near maximum lod
}
\description{
Get SNP information from SQLite database within \code{window_Mbp} of \code{peak_Mbp} on \code{chri_id}
}
\examples{
\dontrun{get_snpprobs(chr_id, peak_Mbp, window_Mbp, scan_obj, probs_obj, datapath)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
