% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs8.R
\name{get_svs8}
\alias{get_svs8}
\title{Get SVS information in window around peak}
\usage{
get_svs8(chr_id, peak_Mbp, window_Mbp, datapath)
}
\arguments{
\item{chr_id}{chromosome identifier}

\item{peak_Mbp}{position in Mbp of peak}

\item{window_Mbp}{half-width of \code{window} around \code{peak_Mbp}}

\item{datapath}{path to Derived Data}
}
\value{
table of class \code{snpinfo} of top_snps near maximum
}
\description{
Get SVS information from RDS database within \code{window_Mbp} of \code{peak_Mbp} on \code{chri_id}
}
\examples{
\dontrun{get_snpinfo(chr_id, peak_Mbp, window_Mbp)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
