% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mgi_features.R
\name{mgigenes_gz2sql}
\alias{mgigenes_gz2sql}
\title{Convert gzipped file of MGI genes to SQLite database}
\usage{
mgigenes_gz2sql(gene_tbl = "mgi_gene", sql_file = "mgi_db.sqlite",
  gz_file = file.path("ftp://ftp.jax.org", "SNPtools", "genes",
  "MGI.20130703.sorted.txt.gz"))
}
\arguments{
\item{gene_tbl}{character string of tbl to be used in database}

\item{sql_file}{character string name of SQLite file}

\item{gz_file}{character string path to MGI gene file}
}
\value{
tbl collected from SQLite database (invisible)
}
\description{
Read file, filter to MGI identified genes and copy to database.
If database does not exist, create it.
The gzipped gene file is 24.8Mb,or 368.2Mb uncompressed.
The R object with only type=="gene" and known Name is 10.5Mb.
Storing in SQLite, file is 1.5Mb.
}
\examples{
dontrun(mgigenes_gz2sql("mgi_gene"))

}
