% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpinfo.R
\name{summary.snpinfo}
\alias{summary.snpinfo}
\title{Summary of snpinfo object}
\usage{
\method{summary}{snpinfo}(object, ...)
}
\arguments{
\item{object}{of class \code{near_snps}}

\item{...}{other arguments not used}
}
\value{
table summary
}
\description{
Summary of snpinfo object
}
\examples{
\dontrun{summary(object)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
