% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_snp.R
\name{summary.feature_snp}
\alias{summary.feature_snp}
\title{Summary of features with SNP information}
\usage{
\method{summary}{feature_snp}(object, ...)
}
\arguments{
\item{object}{tbl of feature information from \code{\link{get_feature_snp}}}

\item{...}{additional parameters ignored}
}
\value{
tbl of feature summaries by type
}
\description{
Summary of features with SNP information
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
