% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exon.R
\name{get_gene_exon_snp}
\alias{get_gene_exon_snp}
\alias{get_gene_exon}
\alias{plot.gene_exon}
\alias{summary.gene_exon}
\alias{subset.gene_exon}
\alias{plot_gene_exon}
\title{Get exons for set of genes}
\usage{
get_gene_exon_snp(top_snps_tbl, sql_filename)

get_gene_exon(feature_tbl, gene_snp)

\method{plot}{gene_exon}(gene_exon, top_snps_tbl = NULL, plot_now = TRUE,
  genes = unique(gene_exon$gene), ...)

\method{summary}{gene_exon}(gene_exon, gene_name = NULL,
  top_snps_tbl = NULL, extra_bp = 5000)

\method{subset}{gene_exon}(x, gene_val, ...)

plot_gene_exon(gene_exon_tbl, top_snps_tbl, gene_name, pheno)
}
\arguments{
\item{top_snps_tbl}{tbl from \code{\link{get_top_snps}} or \code{\link[qtl2scan]{top_snps}}}

\item{sql_filename}{path to \code{\link{get_mgi_features}}}

\item{feature_tbl}{tbl of features from \code{\link{get_mgi_feature}}}

\item{gene_snp}{tbl of genes with SNPs IDs from \code{\link{match_feature_snp}}}

\item{gene_exon}{tbl of feature information from \code{\link{get_gene_exon}}}

\item{plot_now}{plot now if TRUE}

\item{...}{arguments passed along to \code{\link{gene_plot}}}

\item{gene_name}{name of gene as character string}

\item{extra_bp}{extra region beyond gene for SNPs (in bp)}

\item{exon_tbl}{tbl of feature information from \code{\link{get_mgi_features}}}

\item{top_snps_tbl}{tbl of SNP data from \code{\link[qtl2scan]{top_snps}}}

\item{top_snps_tbl}{table of top SNPs in region from \code{\link[qtl2scan]{top_snps}}}
}
\value{
tbl of exon and gene features

tbl of exon and gene features

list of ggplots (see \code{\link{gene_plot}})

tbl of summary
}
\description{
Match up exon start,stop,strand with genes.

Match up exon start,stop,strand with genes.

Uses \code{\link{gene_plot}} to plot genes, exons, mRNA with SNPs.

Returns table of gene and its exons.
}
\examples{
\dontrun{get_gene_exon_snp(top_snps_tbl, sql_filename)}

\dontrun{get_gene_exon(feature_snp)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
\keyword{utilities}
